requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
exceptiongroup==1.1.3
execnet==2.0.2
Faker==20.0.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.9
pycparser==2.21
PySocks==1.7.1
pytest==7.4.3
pytest-lazy-fixture==0.6.3
pytest-parallel==0.1.1
pytest-rerunfailures==13.0
pytest-xdist==3.4.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
selenium==4.15.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tblib==3.0.0
tomli==2.0.1
trio==0.23.1
trio-websocket==0.11.1
urllib3==2.1.0
webdriver-manager==4.0.1
wsproto==1.2.0
